refact: better organize project structure
