+++ constants/standards.py
@@ -1,0 +1,9 @@
+PROJECT_NAME = "NFF"
+PROJECT_FULL_NAME = "NFF - Nota Fiscal Fácil"
+
+TRUTHY_STRS = ["sim", "sin", "si", "s", "1"]
+FALSY_STRS = ["não", "nao", "na", "n", "0"]
+
+STANDARD_SLEEP_TIME = 0.25
+
+BRAZILIAN_DATE_FORMAT = '%d/%m/%Y'
