+++ modules/siare.py
@@ -5,9 +5,10 @@ from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.common.by import By
 from selenium.webdriver.common.keys import Keys
 
+from constants.paths import Urls, XPaths
+from constants.standards import STANDARD_SLEEP_TIME
 from models.entity import Entity
 from models.invoice import Invoice, InvoiceItem
-from utils.constants import STANDARD_SLEEP_TIME, Urls, XPaths
 from utils.helpers import normalize_text
 
 from .browser import Browser
